Split fcl_colours.cpp into multiple files
